+++ TrollCode.Nordnet.API/Responses/LoginResponse.cs
@@ -7,7 +7,7 @@ namespace TrollCode.Nordnet.API.Responses
     public class LoginResponse
     {
         public string Session_key { get; set; }
-        public int Expires_in { get; set; }
+        public long Expires_in { get; set; }
         public string Environment { get; set; }
         public string Country { get; set; }
         public FeedInformation Private_feed { get; set; }
@@ -17,7 +17,7 @@ namespace TrollCode.Nordnet.API.Responses
     public class FeedInformation
     {
         public string Hostname { get; set; }
-        public int Port { get; set; }
+        public long Port { get; set; }
         public bool Encrypted { get; set; }
     }
 }
